refactor($multiple): 시도 횟수 자료형 long 에서 int 로 변경
- Long 사용 시, 메모리 초과 문제가 발생할 것으로 예상.

+++ src/main/java/racingcar/Application.java
@@ -17,7 +17,7 @@ public class Application {
         String carNames = getCarNames();
         List<String> carNameList = getCarNameList(carNames);
         RacingCars racingCars = createRacingCars(carNameList);
-        long tryCount = getTryCount();
+        int tryCount = getTryCount();
     }
 
     private static RacingCars createRacingCars(List<String> carNameList) {
@@ -33,7 +33,7 @@ public class Application {
         return inputView.getCarNames();
     }
 
-    private static long getTryCount() {
+    private static int getTryCount() {
         outputView.printInputTryCount();
         return inputView.getTryCount();
     }
